Extract header component for the sidebar.
This will trigger search.
